Packet implementation + World Gen Progress
Implemented some new packets.
Some more progress on the world generator.
You will now spawn on a bedrock platform :)

+++ SharpMC-Rewrite/Networking/Packets/SpawnPosition.cs
@@ -26,7 +26,7 @@
         public void Write(ClientWrapper state, MSGBuffer buffer, object[] Arguments)
         {
             Vector3 D = Globals.WorldGen.GetSpawnPoint ();
-            long Data = (long)(((int)D.X & 0x3FFFFFF) << 38) | (((int)D.Y & 0xFFF) << 26) | ((int)D.Z & 0x3FFFFFF);
+            long Data = (((long)D.X & 0x3FFFFFF) << 38) | (((long)D.Y & 0xFFF) << 26) | ((long)D.Z & 0x3FFFFFF);
             buffer.WriteVarInt (PacketID);
             buffer.WriteLong (Data);
             buffer.FlushData ();
